setup commands and header
